Put vertical lines where dosing occurs (re-exported from rxode2)

This is a dosing geom that shows the vertical lines where a dose occurs

Value

This returns a stat_amt in context of a ggplot2 plot

Details

Requires the following aesthetics:

x representing the x values, usually time

amt representing the dosing values; They are missing or zero when no dose is given

Examples

library(rxode2)
#> rxode2 1.0.0 using 1 threads (see ?getRxThreads)
#>   no cache: create with `rxCreateCache()`
library(units)
#> Error in library(units): there is no package called ‘units’

## Model from RxODE tutorial
mod1 <-rxode2({
       KA=2.94E-01
       CL=1.86E+01
       V2=4.02E+01
       Q=1.05E+01
       V3=2.97E+02
       Kin=1
       Kout=1
       EC50=200
       C2 = centr/V2
       C3 = peri/V3
       d/dt(depot) =-KA*depot
       d/dt(centr) = KA*depot - CL*C2 - Q*C2 + Q*C3
       d/dt(peri)  =                    Q*C2 - Q*C3
       d/dt(eff)  = Kin - Kout*(1-C2/(EC50+C2))*eff
})
#>  

## These are making the more complex regimens of the rxode2 tutorial

## bid for 5 days
bid <- et(timeUnits="hr") %>%
      et(amt=10000,ii=12,until=set_units(5, "days"))
#> Error in set_units(5, "days"): could not find function "set_units"

## qd for 5 days
qd <- et(timeUnits="hr") %>%
      et(amt=20000,ii=24,until=set_units(5, "days"))
#> Error in set_units(5, "days"): could not find function "set_units"

## bid for 5 days followed by qd for 5 days

et <- seq(bid,qd) %>% et(seq(0,11*24,length.out=100));
#> Error in seq(bid, qd): object 'bid' not found

bidQd <- rxSolve(mod1, et, addDosing=TRUE)
#> Error in rxSolveSEXP(object, .ctl, .nms, .xtra, params, events, inits,     setupOnlyS = .setupOnly): cannot solve without event information

# by default dotted and under-stated
plot(bidQd, C2) + geom_amt(aes(amt=amt))
#> Error in plot(bidQd, C2): object 'bidQd' not found

# of course you can make it a bit more visible

plot(bidQd, C2) + geom_amt(aes(amt=amt), col="red", lty=1, size=1.2)
#> Error in plot(bidQd, C2): object 'bidQd' not found